ci: remove ubuntu-20.04 runner builds on GHA
The Ubuntu 20.04 image on GitHub actions will be unavailable by
2025-04-01, and that is no April Fool's joke. See
<https://github.com/actions/runner-images/issues/11101> for more
information on the deprecation and removal. (Looks like the date
was moved two weeks back to 2025-04-15, but it's time to get rid
of those builds anyway.)

Therefore, the GCC 8 build on Ubuntu 20.04 is removed, as well as
the Clang 7, 8, 9 and 10 builds on Ubuntu 20.04. The software can
still be built on Ubuntu 20.04 as of now, but I can not make any
guarantees that it will stay so in the future.